Moved GNUmakefile to dedicated location, added alternative scripts
Having this GNUmakefile hack was probably not the best idea in first
place since it confuses some users and build systems, for very little
benefit. The main things it did for users can be better be handled in
shell scripts.

INSTALL documents the new workflow and adds some remarks for upcoming
install code integration in cmake scripts.