Printing public name in maintenance logs
For Debian bug #825640, use official hostname to connect to server which
is supposed to make it resolve the name in links to the official
hostname instead of localhost.

Also sending only a mere link to the admin page and not one which
immediately triggers the action, letting admin to maybe change options
(like force getting of index files).