Fix overoptimistic guessing of the SHA256SUMS file location
For Debian bug #928957: incorrect assumption of an existing SHA256SUMS
file for Debian repositories makes the expiration task fail
without a proper way for the end user to recover from it. Now ignore
a download error in this case (similar handling as for other guesses),
assuming that permanent 404ing for other reasons than removal of remote
content can be considered unlikely.